refactor: molecules
