+++ src/components/Molecules/Header/index.tsx
@@ -1,8 +1,6 @@
 import React from "react";
 import { Component } from "./style";
 
-const Header: React.FC<any> = ({ component = "header", ...props }) => {
+export const Header: React.FC<any> = ({ component = "header", ...props }) => {
    return <Component as={component} {...props} />;
 };
-
-export default Header;
