+++ src/components/Molecules/index.ts
@@ -1,0 +1,2 @@
+export * from './Header';
+export * from './Footer';
